+++ src/resolver.ts
@@ -2,8 +2,47 @@ import { Logger } from "ts-log";
 import { resolve as resolvePath } from "path";
 import globby from "globby";
 
+export interface IVirtualPathMapping {
+    /**
+     * Path prefix to match (relative to cwd).
+     */
+    match: string;
+
+    /**
+     * Value to replace matched portion of path with.
+     */
+    replace: string;
+}
+
+/**
+ * A file path mapping.
+ */
+export interface IMappedPath {
+    /**
+     * File's true path.
+     */
+    actual: string;
+
+    /**
+     * File's virtual path.
+     */
+    virtual: string;
+}
+
 interface IResolverOptions {
-    vPathMap: [string, string][];
+    /**
+     * Virtual path mappings.
+     */
+    virtPathMaps: IVirtualPathMapping[];
+
+    /**
+     * Current working directory.
+     */
+    cwd: string;
+
+    /**
+     * Logger.
+     */
     logger: Logger;
 }
 
@@ -15,14 +54,15 @@ interface IResolverOptions {
  *
  * @returns Map of virtual paths to actual paths, all fully resolved to absolute paths.
  */
-export default function (globs: string[], options: IResolverOptions): [ string, string ][] {
+export default function (globs: string[], options: IResolverOptions): IMappedPath[] {
     const log = options.logger;
+    const cwd = options.cwd;
 
     // Resolve all file paths
 
     log.trace("Looking for files", { globs: globs });
 
-    const paths = new Set<string>(globby.sync(globs));
+    const paths = new Set<string>(globby.sync(globs, { cwd }));
 
     log.trace(`Found ${paths.size} files, filtering with provided virtual paths`);
 
@@ -34,28 +74,28 @@ export default function (globs: string[], options: IResolverOptions): [ string, 
      * - 1 is the real path
      * - 2 is the file perference
      */
-    const candidatePaths = new Map<string, [string, number]>();
+    const candidatePaths = new Map<string, { actual: string, preference: number}>();
 
-    for (const path of paths) {
-        log.trace("Inspecting file path", { path });
+    for (const actual of paths) {
+        log.trace("Inspecting file path", { path: actual });
 
         // Resolve path
-        const [ vPath, preference ] = resolveVirtualPath(path, options.vPathMap, log);
+        const [ virtual, preference ] = resolveVirtualPath(actual, options.virtPathMaps, cwd, log);
 
         // Add to resolved file if highest preference
-        const existingFile = candidatePaths.get(vPath);
+        const existingFile = candidatePaths.get(virtual);
         if (existingFile) {
-            if (existingFile[1] < preference) {
-                log.trace("Superseeding existing file path", { vPath, path, preference });
-                candidatePaths.set(vPath, [ path, preference ]);
+            if (existingFile.preference < preference) {
+                log.trace("Superseeding existing file path", { virtual, actual, preference });
+                candidatePaths.set(virtual, { actual, preference });
             }
             else {
-                log.trace("Discarding new file path due to lower preference", { vPath, path, preference });
+                log.trace("Discarding new file path due to lower preference", { virtual, actual, preference });
             }
         }
         else {
-            log.trace("Adding file path", { vPath, path, preference });
-            candidatePaths.set(vPath, [ path, preference ]);
+            log.trace("Adding file path", { virtual, actual, preference });
+            candidatePaths.set(virtual, { actual, preference });
         }
     }
 
@@ -63,10 +103,10 @@ export default function (globs: string[], options: IResolverOptions): [ string, 
 
     // Simplify outputs
 
-    const resolvedPaths: [ string, string ][] = [];
+    const resolvedPaths: IMappedPath[] = [];
 
-    for (const [ vPath, [ path ] ] of candidatePaths.entries()) {
-        resolvedPaths.push([ resolvePath(vPath), resolvePath(path) ]);
+    for (const [ virtual, { actual } ] of candidatePaths.entries()) {
+        resolvedPaths.push({ virtual, actual: resolvePath(cwd, actual) });
     }
 
     return resolvedPaths;
@@ -75,29 +115,34 @@ export default function (globs: string[], options: IResolverOptions): [ string, 
 /**
  * Attempts to resolve path to a virtual path, returning provided path on failure.
  *
- * @param path Absolute path to try and resolve.
- * @param vPaths Virtual path mappings.
- * @param log Logger
+ * @param actual - Absolute path to try and resolve.
+ * @param virtPathMaps - Virtual path mappings.
+ * @param log - Logger
  *
  * @returns New or existing path and preference.
  */
-function resolveVirtualPath(path: string, vPaths: [string, string][], log: Logger): [string, number] {
+function resolveVirtualPath(
+    actual: string,
+    virtPathMaps: IVirtualPathMapping[],
+    cwd: string,
+    log: Logger
+): [string, number] {
     // Try to resolve a virtual path
-    log.trace("Attempting to resolve virtual path", { path });
+    log.trace("Attempting to resolve virtual path", { path: actual });
 
     let preference = 0;
-    for (const [ oldPathStart, newPathStart ] of vPaths) {
-        if (path.startsWith(oldPathStart)) {
+    for (const { match, replace } of virtPathMaps) {
+        if (actual.startsWith(match)) {
             // Virtual path match discovered
-            const resolvedPath = resolvePath(path.replace(oldPathStart, newPathStart));
-            log.trace("Resolved virtual path", { path, resolvedPath, preference });
-            return [ resolvedPath, preference ];
+            const virtual = resolvePath(cwd, actual.replace(match, replace));
+            log.trace("Resolved virtual path", { actual, virtual, preference });
+            return [ virtual, preference ];
         }
 
         preference++;
     }
 
     // No matches, lowest preference
-    log.trace("No matching virtual path mappings", { path, preference: 0 });
-    return [ path, 0] ;
+    log.trace("No matching virtual path mappings", { actual, preference: 0 });
+    return [ resolvePath(cwd, actual), 0] ;
 }